requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi>=0.116.1
uvicorn>=0.35.0
jinja2>=3.1.5
python-multipart>=0.0.18
itsdangerous>=2.1.2          # required by starlette SessionMiddleware

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.8         # required by pymysql for TiDB SSL connections

# ── AI / LLM ─────────────────────────────────────────────────────────────────
groq>=0.9.0                  # triage (LLaMA 3) + audio transcription (Whisper)
langchain-google-genai>=1.0.7 # Gemini 2.5 Flash for discharge extraction
langchain-core>=0.3.0        # HumanMessage and base LangChain types

# ── Utilities ────────────────────────────────────────────────────────────────
pydantic==2.8.2
python-dotenv==1.0.1
reportlab>=4.0.0             # PDF discharge report generation
